swagger: rename 'docs/' to 'swagger/'
Use 'docs/' for zot-related documentation.
