+++ swagger/docs.go
@@ -2,7 +2,7 @@
 // This file was generated by swaggo/swag at
 // 2019-12-11 12:03:05.055900322 -0800 PST m=+0.052058015
 
-package docs
+package swagger
 
 import (
 	"bytes"
