+++ swagger/docs_test.go
@@ -1,15 +1,15 @@
-package docs_test
+package swagger_test
 
 import (
 	"testing"
 
-	"github.com/anuvu/zot/docs"
+	"github.com/anuvu/zot/swagger"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
 func TestDocs(t *testing.T) {
 	Convey("Read docs", t, func() {
-		s := docs.New()
+		s := swagger.New()
 		So(s.ReadDoc(), ShouldNotBeEmpty)
 	})
 }
